frontend/requirements.txt
streamlit==1.28.0
requests==2.31.0
pandas==2.1.2
numpy==1.26.1
matplotlib==3.8.1
seaborn==0.13.0
pillow==10.1.0
plotly==5.18.0
python-dotenv==1.0.0
watchdog==3.0.0